+++ Invoke-InstallModule.ps1
@@ -32,7 +32,7 @@ Begin {
             'CyGlobalLists.ps1',
             'CyUsers.ps1',
             'CyOptics.ps1',
-            'License.txt'
+            'license.txt'
         )
     }
     Catch {